+++ packages/server/src/unstable-core-do-not-import/procedure.ts
@@ -1,7 +1,4 @@
-import type { inferObservableValue } from '../observable';
-import type { ClientInferrable, inferRootTypes } from './clientish/inferrable';
 import type { ProcedureCallOptions } from './procedureBuilder';
-import type { Serialize } from './serialize';
 
 export const procedureTypes = ['query', 'mutation', 'subscription'] as const;
 /**
@@ -64,33 +61,3 @@ export type AnyQueryProcedure = QueryProcedure<any>;
 export type AnyMutationProcedure = MutationProcedure<any>;
 export type AnySubscriptionProcedure = SubscriptionProcedure<any>;
 export type AnyProcedure = Procedure<ProcedureType, any>;
-
-/**
- * @internal
- * TODO: move to `@trpc/client`
- */
-export type inferTransformedProcedureOutput<
-  TInferrable extends ClientInferrable,
-  TProcedure extends AnyProcedure,
-> = inferRootTypes<TInferrable>['transformer'] extends false
-  ? Serialize<TProcedure['_def']['_output_out']>
-  : TProcedure['_def']['_output_out'];
-
-/** @internal */
-export type inferTransformedSubscriptionOutput<
-  TInferrable extends ClientInferrable,
-  TProcedure extends AnyProcedure,
-> = inferRootTypes<TInferrable>['transformer'] extends false
-  ? Serialize<inferObservableValue<TProcedure['_def']['_output_out']>>
-  : inferObservableValue<TProcedure['_def']['_output_out']>;
-
-export type inferProcedureInput<TProcedure extends AnyProcedure> =
-  undefined extends inferProcedureParams<TProcedure>['_input_in']
-    ? void | inferProcedureParams<TProcedure>['_input_in']
-    : inferProcedureParams<TProcedure>['_input_in'];
-
-export type inferProcedureParams<TProcedure> = TProcedure extends AnyProcedure
-  ? TProcedure['_def']
-  : never;
-export type inferProcedureOutput<TProcedure> =
-  inferProcedureParams<TProcedure>['_output_out'];
